+++ static/js/admin.ts
@@ -1,5 +1,3 @@
-import { autoSave } from "./autosave";
-
 export interface InitializeAdminUsersPageOptions {
   readonly page: 'admin-users';
 }
@@ -36,7 +34,35 @@ export function initializeAdminUserPage(_options: InitializeAdminUsersPageOption
       $('#hidden_page_input').attr('value', token);
       $('#filterform').submit();
   });
+}
 
-    // Autosave admin page
-    autoSave("admin_users")
+export function filter_admin() {
+    const params: Record<string, any> = {};
+  
+    const filter = $('#admin_filter_category').val();
+    params['filter'] = filter;
+  
+    if ($('#hidden_page_input').val()) {
+      params['page'] = $('#hidden_page_input').val();
+    }
+  
+    switch (filter) {
+      case 'email':
+      case 'username':
+        params['substring'] = $('#email_filter_input').val();
+        break;
+      case 'language':
+        params['language'] = $('#language_filter_input').val();
+        break;
+      case 'keyword_language':
+        params['keyword_language'] = $('#keyword_language_filter_input').val();
+        break;
+      default:
+        params['start'] = $('#admin_start_date').val();
+        params['end'] = $('#admin_end_date').val();
+        break;
+    }
+  
+    const queryString = Object.entries(params).map(([k, v]) => k + '=' + encodeURIComponent(v)).join('&');
+    window.open('?' + queryString, '_self');
 }
