+++ static/js/index.ts
@@ -13,6 +13,7 @@
  */
 import './htmx-integration';
 export * from './modal';
+export * from './admin';
 export * from './app';
 export * from './auth';
 export * from './statistics';
