+++ templates/admin/admin-users.html
@@ -167,11 +167,27 @@
             <td class="experience_cell hidden">{{user.prog_experience}}</td>
             <td class="tags_cell text-center cursor-pointer"><span onclick="hedyApp.edit_user_tags('{{user.username}}');">✏️</span></td>
             {% if user.teacher_request %}
-                <td class="teacher_cell text-center"><span class="fa-solid fa-circle-question cursor-pointer" onclick="hedyApp.markAsTeacher (this, '{{user.username}}', true, true)"></span></td>
-            {% else %}
-                <td class="teacher_cell text-center"><input type="checkbox" {% if user.is_teacher %}checked="checked"{% endif %} onclick="hedyApp.markAsTeacher (this, '{{user.username}}', {% if user.is_teacher %}false{% else %}true{% endif %}, false)"></td>
+                <td id="teacher_request_{{user.username}}" class="teacher_cell text-center">
+                    <span class="fa-solid fa-circle-question cursor-pointer"
+                    hx-trigger="click"
+                    hx-post="/admin/mark-as-teacher/{{user.username}}/{{user.is_teacher}}"
+                    _="on click
+                    remove #teacher_request_{{user.username}}
+                    then toggle .hidden on #is_teacher_{{user.username}}
+                    then set #check_teacher_{{user.username}}.checked to true">
+                    </span>
+                </td>
             {% endif %}
-            <td class="teacher_cell text-center"><input type="checkbox" {% if user.is_super_teacher %}checked="checked"{% endif %} onclick="hedyApp.markSuperTeacher (this, '{{user.username}}', {% if user.is_super_teacher %}true{% else %}false{% endif %})"></td>
+                <td id="is_teacher_{{user.username}}" class="{% if user.teacher_request %}hidden{% endif %} teacher_cell text-center">
+                    <input id="check_teacher_{{user.username}}" type="checkbox" {% if user.is_teacher %}checked{% endif %}
+                    hx-trigger="click"
+                    hx-post="/admin/mark-as-teacher/{{user.username}}/{{user.is_teacher}}">
+                </td>
+            <td class="teacher_cell text-center">
+                <input type="checkbox" {% if user.is_super_teacher %}checked="checked"{% endif %} 
+                hx-trigger="click"
+                hx-post="/admin/mark-super-teacher/{{user.username}}">
+            </td>
             <td class="verified_cell hidden">{{user.email_verified}}</td>
             <td class="third_party_cell hidden text-center"><input type="checkbox" {% if user.third_party %}checked="checked"{% endif %} ></td>
           </tr>
